Вроде заработали транзацкии

+++ src/main/java/models/interfaces/IModel.java
@@ -2,6 +2,6 @@ package models.interfaces;
 
 public interface IModel {
 
-    public Integer getId();
+    public Long getId();
 
 }
